// frontend/app/controller/model/workflow/workflowPresetDTO.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { StepPresetDTO } from "./stepPresetDTO";

export interface WorkflowPresetDTO {
    id: number;
    title: string;
    description?: string;
    stepPresets?: StepPresetDTO[];
    dateC?: string;
    dateUpdate?: string;
    createurId?: number;
    createurNom?: string;
    createurPrenom?: string;
    departement?: string;

    

  }